Change heuristic function to return float

+++ Jansk.Pathfinding.Tests/BasePathfindingTest.cs
@@ -5,7 +5,7 @@ namespace Jansk.Pathfinding.Tests
 {
     public class BasePathfindingTest
     {
-        protected readonly Func<Tile, Tile, int> _heuristic = (from, to) =>
+        protected readonly Func<Tile, Tile, float> _heuristic = (from, to) =>
             Math.Abs(from.x - to.x) + Math.Abs(from.y - to.y) + Math.Abs(from.z - to.z);
     }
 }
